[Bugfix]  asyncscheduler bug fix
